\name{BayesBDshiny}
\alias{BayesBDshiny}
\title{
Boundary detection and shiny integration
}
\description{
This function produces a local shiny instance for easily producing graphics of boundary detection simulations and real-data analyses.
}
\usage{
BayesBDshiny()
}
\examples{
BayesBDshiny()
}
