\name{example_image_data}
\alias{example_image_data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Example data
}
\description{
This is an example of image data for a Gaussian-noised image.  This is the correct format for data uploads when using BayesBDshiny().
}

