\name{BayesBD-package}
\alias{BayesBD-package}
\alias{BayesBD}
\docType{package}
\title{
  Bayesian Inference for Image Boundaries
}
\description{
  This package provides tools for carrying out a Bayesian analysis of image boundaries.  Functions are provided
  for both binary (Bernoulli) and continuous (Gaussian) images.  Examples, along with an interactive shiny function
  illustrate how to perform simulations, analyze custom data, and plot estimates and credible intervals.    
}
\details{
Package: BayesBD
License: GPL (>= 3) 
Imports: Rcpp (>= 0.12.5), RcppArmadillo, shiny, png, jpeg, mritc
LinkingTo: Rcpp, RcppArmadillo
}
\author{
Nicholas Syring, Meng Li

Maintainer: Nicholas Syring <nasyrin@gmail.com>
}
\references{
Li, M. and Ghosal, S.(2015) "Bayesian Detection of Image Boundaries." arXiv 1508.05847.
}

